refactor: 🛠️ update Axios methods and add apis/utils directories
